requirements.txt
streamlit==1.25.0
pandas==2.0.3
requests==2.31.0
openai==0.28.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
